+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -236,5 +236,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-08T14:08:02.887Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-08T21:44:52.961Z")
     r"""The date and time the resource was last updated"""