+++ packages/orq-rc/RELEASES.md
@@ -3219,3 +3219,13 @@ Based on:
 - [python v3.14.0-rc.17] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.17] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.17 - packages/orq-rc
+
+## 2025-10-08 21:46:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.2 (2.723.8) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.18] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.18] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.18 - packages/orq-rc